+++ src/langs/javascript.mjs
@@ -76,7 +76,7 @@ export default {
 	},
 	string_single: {
 		end: {
-			regex: /"/g,
+			regex: /'/g,
 			switch_state: "default"
 		}
 	},