+++ tests/javascript/quotes-backtick.in.js
@@ -1,0 +1,7 @@
+let item = `orange`;
+
+function do_work() {
+	return 2 + 2;
+}
+
+console.log(do_work());