+++ CHANGELOG.md
@@ -17,6 +17,7 @@ npm install --global skyliner # For using the CLI globally
 
 
 ## v0.3 (unreleased)
+ - javascript: Fix single quotes
  - Ensure that earlier Lexer sessions are not resumed after a new Lexer session begins
      - If you need to outline multiple things at once with `Skyliner.outline_iterate()`, it is advised you create multiple `Skyliner` class instances.